Pruneu command

+++ cogs/moderation.py
@@ -1,5 +1,4 @@
-import asyncio
-
+import discord
 from discord.ext import commands
 
 from utils import checks
@@ -9,9 +8,21 @@ class Moderation(commands.Cog):
 	@commands.command()
 	@checks.privileged()
 	async def prune(self, ctx, how_many: int) -> None:
-		log = await ctx.send(f"_💣ing {how_many} messages_")
-		async for message in ctx.channel.history(limit=how_many, before=log):
+		if how_many > 50:
+			await ctx.send("You can delete up to 50 messages")
+			return
+		async for message in ctx.channel.history(limit=how_many):
 			await message.delete()
-		await log.edit(content="👌")
-		await asyncio.sleep(2)
-		await log.delete()
+
+	@commands.command()
+	@checks.privileged()
+	async def pruneu(self, ctx, member: discord.User, how_many: int) -> None:
+		if how_many > 50:
+			await ctx.send("You can delete up to 50 messages")
+			return
+		c = 0
+		async for message in ctx.channel.history(limit=2 * how_many):
+			if c >= how_many or message.author != member:
+				continue
+			await message.delete()
+			c += 1
